Bump Pandoc dependency
Stack lts 8.0 ships with Pandoc 1.19.2.1, which is outside our existing
dependency graph. That means that end users that want to use this
package need to define a number of additional dependencies for their
stack projects if they wish (or need) to use 8.0.

Bumping the dependency here will allow users to simply add
yesod-markdown as a dependency, without needing to change the version of
Pandoc they use.