Stop running weeder
Weeder-2.0 removes the extra dependencies feature (and improves the dead
code elimination feature). In a library, the new dead-code behavior
isn't great, since you tend to have wide public interface of "roots". We
were using Weeder-1.0 mostly for the extra dependencies feature, which
is now part of GHC (or Cabal) as of 8.10.